refactor: move utils to domain/validators
