+++ ProtoCredentials/OpenID4VC-Prototype/Application/Services/VerifierService.cs
@@ -3,7 +3,7 @@ using OpenID4VC_Prototype.Application.Interfaces;
 using OpenID4VC_Prototype.Application.Models;
 using OpenID4VC_Prototype.Domain.Interfaces;
 using OpenID4VC_Prototype.Domain.Models;
-using OpenID4VC_Prototype.Utils;
+using OpenID4VC_Prototype.Domain.Validators;
 
 namespace OpenID4VC_Prototype.Application.Services;
 
@@ -15,7 +15,7 @@ public class VerifierService(ICryptoService cryptoService) : IVerifierService
 
         var domainCredential = credential.Adapt<VerifiableCredential>();
 
-        if (!CredentialUtils.IsValidCredential(domainCredential))
+        if (!CredentialValidators.IsValidCredential(domainCredential))
             return new ValidationResult(false, "Invalid Credential provided");
 
         if (string.IsNullOrEmpty(issuerPublicKey))
