+++ ProtoCredentials/OpenID4VC-Prototype/Domain/Validators/DIdValidators.cs
@@ -1,8 +1,8 @@
 ﻿using System.Text.RegularExpressions;
 
-namespace OpenID4VC_Prototype.Utils;
+namespace OpenID4VC_Prototype.Domain.Validators;
 
-public static class DIdUtils
+public static class DIdValidators
 {
     public static bool IsValidDId(string dId)
     {
